+++ modules/Grades/GPARankList.php
@@ -11,6 +11,8 @@ if ( empty( $_REQUEST['mp'] ) )
 }
 
 // Get all the mp's associated with the current mp
+// Fix PostgreSQL error invalid ORDER BY, only result column names can be used
+// Do not use ORDER BY SORT_ORDER IS NULL,SORT_ORDER (nulls last) here.
 $mps_RET = DBGet( "SELECT MARKING_PERIOD_ID,TITLE,DOES_GRADES,0,SORT_ORDER
 FROM school_marking_periods
 WHERE MARKING_PERIOD_ID=(SELECT PARENT_ID FROM school_marking_periods WHERE MARKING_PERIOD_ID=(SELECT PARENT_ID FROM school_marking_periods WHERE MARKING_PERIOD_ID='" . UserMP() . "'))
@@ -29,7 +31,7 @@ SELECT MARKING_PERIOD_ID,TITLE,DOES_GRADES,3,SORT_ORDER
 FROM school_marking_periods
 WHERE PARENT_ID='" . UserMP() . "'
 AND MP='PRO'
-ORDER BY 5,SORT_ORDER IS NULL,SORT_ORDER" );
+ORDER BY 5,SORT_ORDER" );
 
 if ( $_REQUEST['search_modfunc'] === 'list' )
 {