+++ modules/Students/includes/Address.inc.php
@@ -430,6 +430,8 @@ echo ErrorMessage( $error );
 
 if ( ! $_REQUEST['modfunc'] )
 {
+	// Fix PostgreSQL error invalid ORDER BY, only result column names can be used
+	// Do not use ORDER BY SORT_ORDER IS NULL,SORT_ORDER (nulls last) here.
 	$addresses_RET = DBGet( "SELECT a.ADDRESS_ID, sjp.STUDENT_RELATION,a.ADDRESS,
 		a.CITY,a.STATE,a.ZIPCODE,a.PHONE,a.MAIL_ADDRESS,a.MAIL_CITY,a.MAIL_STATE,a.MAIL_ZIPCODE,
 		sjp.CUSTODY,sja.MAILING,sja.RESIDENCE,sja.BUS_PICKUP,sja.BUS_DROPOFF," .
@@ -451,7 +453,7 @@ if ( ! $_REQUEST['modfunc'] )
 		FROM students_join_people sjp
 		WHERE sjp.STUDENT_ID=sja.STUDENT_ID
 		AND sjp.ADDRESS_ID=a.ADDRESS_ID)
-	ORDER BY SORT_ORDER IS NULL,SORT_ORDER,RESIDENCE,CUSTODY,STUDENT_RELATION", [], [ 'ADDRESS_ID' ] );
+	ORDER BY SORT_ORDER,RESIDENCE,CUSTODY,STUDENT_RELATION", [], [ 'ADDRESS_ID' ] );
 
 	//echo '<pre>'; var_dump($addresses_RET); echo '</pre>';