+++ CHANGES.md
@@ -62,7 +62,7 @@ Changes in 10.0
 - Add MySQLRemoveDelimiter() remove DELIMITER $$ declarations before procedures or functions in database.inc.php, Modules.inc.php & Plugins.inc.php
 - SQL ORDER BY SORT_ORDER IS NULL,SORT_ORDER (nulls last) for consistency between PostgreSQL & MySQL, program wide
 - HTML use number input for Gradebook config options in Configuration.php
-- Fix PostgreSQL error invalid ORDER BY, only result column names can be used in Portal.php & TakeAttendance.php
+- Fix PostgreSQL error invalid ORDER BY, only result column names can be used, program wide
 
 Changes in 9.3
 --------------